+++ src/components/cartItems/cartSummary.tsx
@@ -2,11 +2,9 @@
 
 import React from 'react';
 import { useRouter } from 'next/navigation';
-
 import { useAuthStore } from '@/app/auth/useAuthStore';
 import { useOrder } from '@/hooks/order/useOrder';
 import { CartSummaryProps } from '@/types/cart';
-
 import { Button } from '../ui/Button';
 import { showCustomToast } from '../ui/Toast/Toast';
